Replace only the first "s" in a subdomain with "%"
Currently, every "-" is replaced with ":" and every "s" is replaced with
"%". This is a problem on Linux where the scope ID is the textual
interface name, especially since systemd's predictable interface naming
scheme uses "s" to indicate the PCI slot of a card (e.g. enp0s31f6).

To make these scope IDs usable, only substitute the first "s" instead of
all of them. Ideally, we should also only substitute dashes that show up
before the "s", but to my knowledge no OS uses dashes in its scope
names, so this remaining issue should be harmless.